Re-designed the script locations so that they make a bit more sense as to what script belongs to what function.
Seperated out files which are generated at run time from the environmental files by creating a 'run' sub-directory.

Switch from embedded script sourcing to make each script independant of each other.

Seperated out the functions on the undercloud to split the VM creation from the configuration.

Changed the way networks get mapped to Overcloud VM's.  Now you pick a KVM network and assign it to a
specific ethernet port on the VM.  This will allow for a closer configuration to match a customer.

Added some tools that are useful as long as the node names contain their role in them (ie, control, compute, ceph-storage).

Added comments to the enviroment files.

Made the networks more flexible so you're not locked into a /24 anywhere.  You can set the network size on all networks.